pyproject.toml
[tool.poetry]
name = "auto_for_wechat_publishing_v2"
version = "0.1.0"
# Updated version from v1 implied by prompt
description = "Automates publishing Markdown articles to WeChat Official Account drafts."
authors = ["Jimmy <15828061015@163.com>"] 
# Replace with your details
readme = "README.md"
packages = [{include = "auto_for_wechat_publishing", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9" 
# Specify your Python version
requests = "^2.31.0"
markdown = "^3.6"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
beautifulsoup4 = "^4.12.3"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
# Add other dev dependencies like linters (flake8, black) if desired

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
publish-wechat = "auto_for_wechat_publishing.main:run" 
# Example CLI entry point